+++ R/method-ggplot-add.R
@@ -29,6 +29,22 @@ ggplot_add.geom_range <- function(object, plot, object_name) {
     ggplot_add(obj, plot, object_name)
 }
 
+##' @method ggplot_add layout_dendrogram
+##' @export
+ggplot_add.layout_dendrogram <- function(object, plot, object_name) {
+    plot <- revts(plot)
+    obj <- list(scale_x_reverse(labels = abs),
+                coord_flip(clip = 'off')
+                )
+    ggplot_add(obj, plot, object_name)
+}
+
+##' @method ggplot_add layout_fan
+##' @export
+ggplot_add.layout_fan <- function(object, plot, object_name) {
+    open_tree(plot, object$angle)
+}
+
 ##' @method ggplot_add range_xaxis
 ##' @export
 ggplot_add.range_xaxis <- function(object, plot, object_name) {